move addWordModal into own component
