+++ components/MyWordlist.js
@@ -1,55 +1,24 @@
-import { Alert, Button, Form, FormGroup, Input, Label, Modal, ModalBody, ModalHeader } from 'reactstrap';
+import { Alert } from 'reactstrap';
 import { useState } from 'react';
 
 import AddWordIcon from './AddWordIcon';
+import AddWordModal from './AddWordModal';
 import RefreshIcon from './RefreshIcon';
-import ValidationError from '../errors/validation-error';
 import ViewConfigInterface from './ViewConfigInterface';
 import WordlistEntry from './WordlistEntry';
-import { v4 as uuidv4 } from 'uuid';
 
 const rSelectedValues = {
   CATEGORIES: 'CATEGORIES',
   DESCRIPTIONS: 'DESCRIPTIONS'
 };
 
-const wordlistEntry = ({ description, wordName }) => {
-  if (!wordName) throw new ValidationError('wordName cannot be empty');
-
-  return {
-    categories: [],
-    description,
-    id: uuidv4(),
-    word: { name: wordName }
-  };
-};
-
 const MyWordlist = ({ wordlistEntriesData }) => {
   const { CATEGORIES, DESCRIPTIONS } = rSelectedValues;
-  const [addWordModal, setAddWordModal] = useState(false);
+  const [addWordModalIsOpen, setAddWordModalIsOpen] = useState(false);
   const [alertVisible, setAlertVisible] = useState(false);
-  const [description, setDescription] = useState('');
   const [rSelected, setRSelected] = useState(CATEGORIES);
   const [showAddWordIcon, setShowAddWordIcon] = useState(true);
-  const [wordName, setWordName] = useState('');
   const [wordlistEntries, setWordlistEntries] = useState(wordlistEntriesData);
-
-  const handleChange = ({ target: { name, value } }) => {
-    if (name === 'wordName') {
-      return setWordName(value);
-    }
-
-    if (name === 'description') {
-      return setDescription(value);
-    }
-  };
-
-  const handleSubmit = async event => {
-    event.preventDefault();
-    setWordlistEntries([ wordlistEntry({ description, wordName }), ...wordlistEntries ]);
-  };
-
-  const onDismiss = () => setAlertVisible(false);
 
   const renderWordlistEntries = entries => entries.map(({
     categories,
@@ -74,30 +43,24 @@ const MyWordlist = ({ wordlistEntriesData }) => {
     );
   });
 
-  const toggleAddWordModal = () => setAddWordModal(!addWordModal);
+  const toggleAddWordModal = () => setAddWordModalIsOpen(!addWordModalIsOpen);
+
+  const addWordModalProps = {
+    isOpen: addWordModalIsOpen,
+    setWordlistEntries,
+    toggle: toggleAddWordModal,
+    wordlistEntries
+  };
+
   const viewConfigInterfaceProps = { rSelected, rSelectedValues, setRSelected };
+
   return (
     <>
-      <Alert color={'warning'} isOpen={alertVisible} toggle={onDismiss}>
+      <Alert color={'warning'} isOpen={alertVisible} toggle={() => setAlertVisible(false)}>
         Wordlist entry failed to upload. Tap the <RefreshIcon bottom='0.05em' height='0.85em' /> icon to try again.
       </Alert>
       <ViewConfigInterface {...viewConfigInterfaceProps} />
-      <Modal isOpen={addWordModal} toggle={toggleAddWordModal}>
-        <ModalHeader toggle={toggleAddWordModal}>New Word</ModalHeader>
-        <ModalBody>
-          <Form onSubmit={handleSubmit}>
-            <FormGroup>
-              <Label for='wordName' style={{ marginBottom: '0' }}>Word:</Label>
-              <Input aria-label='word' id='name' minLength='2' maxLength='64' name='wordName' onChange={handleChange} type='text' />
-            </FormGroup>
-            <FormGroup>
-              <Label for='description' style={{ marginBottom: '0' }}>Description:</Label>
-              <textarea aria-label='definition' className='form-control' id='description' name='description' onChange={handleChange} />
-            </FormGroup>
-            <Button color='primary' onClick={toggleAddWordModal} type='submit'>Add to Wordlist!</Button>{' '}
-          </Form>
-        </ModalBody>
-      </Modal>
+      <AddWordModal {...addWordModalProps} />
       <ul style={{ listStyle: 'none', padding: '0' }}>
         {renderWordlistEntries(wordlistEntries)}
       </ul>
